feat: add input validation and exception handlers

+++ src/main/java/com/panda/library/service/BookService.java
@@ -1,11 +1,13 @@
 package com.panda.library.service;
 
+import com.panda.library.exception.BookNotFoundException;
 import com.panda.library.model.Book;
 import com.panda.library.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookService {
@@ -17,19 +19,20 @@ public class BookService {
         return bookRepository.findAll();
     }
 
-    public Book getBookById(Long id) {
-        return bookRepository.findById(id).orElse(null);
+    public Book getBookById(Long id) throws BookNotFoundException {
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()) {
+            return book.get();
+        }
+        throw new BookNotFoundException("Book not found");
     }
 
     public Book createBook(Book book) {
         return bookRepository.save(book);
     }
 
-    public Book updateBook(Long id, Book book) {
+    public Book updateBook(Long id, Book book) throws BookNotFoundException {
         Book oldBook = getBookById(id);
-        if (oldBook == null) {
-            return null;
-        }
         if(book.getTitle() != null)
             oldBook.setTitle(book.getTitle());
         if(book.getAuthor() != null)
@@ -41,7 +44,10 @@ public class BookService {
         return bookRepository.save(oldBook);
     }
 
-    public void deleteBook(Long id) {
+    public void deleteBook(Long id) throws BookNotFoundException {
+        if(!bookRepository.existsById(id)){
+            throw new BookNotFoundException("Book not found");
+        }
         bookRepository.deleteById(id);
     }
 }
