+++ src/main/java/com/panda/library/service/PatronService.java
@@ -1,11 +1,13 @@
 package com.panda.library.service;
 
+import com.panda.library.exception.PatronNotFoundException;
 import com.panda.library.model.Patron;
 import com.panda.library.repository.PatronRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class PatronService {
@@ -17,27 +19,31 @@ public class PatronService {
         return patronRepository.findAll();
     }
 
-    public Patron getPatronById(Long id) {
-        return patronRepository.findById(id).orElse(null);
+    public Patron getPatronById(Long id) throws PatronNotFoundException {
+        Optional<Patron> patron = patronRepository.findById(id);
+        if (patron.isPresent()) {
+            return patron.get();
+        }
+        throw new PatronNotFoundException("Patron Not Found");
     }
 
     public Patron createPatron(Patron patron) {
         return patronRepository.save(patron);
     }
 
-    public Patron updatePatron(Long id, Patron patron) {
-        Patron oldPatron = patronRepository.findById(id).orElse(null);
-        if(oldPatron == null){
-            return null;
-        }
+    public Patron updatePatron(Long id, Patron patron) throws PatronNotFoundException {
+        Patron oldPatron = getPatronById(id);
         if(patron.getName() != null)
             oldPatron.setName(patron.getName());
-        if(patron.getContactInformation() != null)
-            oldPatron.setContactInformation(patron.getContactInformation());
+        if(patron.getPhoneNumber() != null)
+            oldPatron.setPhoneNumber(patron.getPhoneNumber());
         return patronRepository.save(oldPatron);
     }
 
-    public void deletePatron(Long id) {
+    public void deletePatron(Long id) throws PatronNotFoundException {
+        if(!patronRepository.existsById(id)){
+            throw new PatronNotFoundException("Patron Not Found");
+        }
         patronRepository.deleteById(id);
     }
 }
